+++ amiss-api/src/test/java/amiss/api/web/UniversityControllerTest.java
@@ -51,7 +51,7 @@ class UniversityControllerTest {
 
     private static SaveState save() {
         return new SaveState(7L, "bob", "My Save", 3, 3, 3960, 3, 70, 0, 0, 0, 1, 1,
-                null, 60, 30, 40, null, 0, 200, 100, 30, 50, false, 0, 0);
+                null, 60, 30, 40, null, 0, 200, 100, 30, 50, false, (byte) 0, (short) 0);
     }
 
     private static SaveStateDto dto() {
